Change the layout of the module
