+++ client/src/app/(modules)/network/(main)/(index)/filters-sidebar.tsx
